# expense-management-app/API.yaml
openapi: 3.0.0
info:
  title: Expense Management API
  version: 1.0.0
  description: Full-stack expense management API with payment processing and AI insights
  contact:
    name: Support
    url: https://github.com/yourusername/expense-management-app
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                first_name:
                  type: string
                last_name:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/profile:
    get:
      summary: Get current user profile
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /expenses:
    get:
      summary: Get all expenses for current user
      tags: [Expenses]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, reimbursed]
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseList'

    post:
      summary: Create new expense with optional receipt
      tags: [Expenses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                amount:
                  type: number
                  format: decimal
                category_id:
                  type: string
                  format: uuid
                expense_date:
                  type: string
                  format: date
                payment_method:
                  type: string
                  enum: [cash, card, check, online]
                notes:
                  type: string
                receipt:
                  type: string
                  format: binary
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
...
  /admin/users:
    get:
      summary: Get all users (admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /admin/statistics/dashboard:
    get:
      summary: Get dashboard statistics (admin only)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [user, admin]
        is_active:
          type: boolean
        is_email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        amount:
          type: number
          format: decimal
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        expense_date:
          type: string
          format: date
        status:
          type: string
          enum: [pending, approved, rejected, reimbursed]
        receipt_url:
          type: string
        payment_method:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ExpenseList:
      type: object
      properties:
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    DashboardStats:
      type: object
      properties:
        statistics:
          type: object
          properties:
            users:
              type: object
              properties:
                total:
                  type: integer
            expenses:
              type: object
              properties:
                total:
                  type: integer
                totalAmount:
                  type: number
                  format: decimal
                pending:
                  type: integer
                pendingAmount:
                  type: number
                  format: decimal
            payments:
              type: object
              properties:
                successful:
                  type: integer
                successfulAmount:
                  type: number
                  format: decimal
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT